+++ mvc/src/main/java/com/interface21/webmvc/servlet/mvc/tobe/adapter/HandlerAdapters.java
@@ -8,7 +8,6 @@ public class HandlerAdapters {
 
     public HandlerAdapters() {
         this.handlerAdapters = List.of(
-                new ControllerHandlerAdapter(),
                 new HandlerExecutionHandlerAdapter()
         );
     }
